// interfaces/Tea.ts
export interface Tea {
  caffeineLevel: any;
  _id?: string;
  name: string;
  slug: string;
  altnames: string;
  origin: string;
  type: string;
  temperature: string;
  brewTime: string;
  image: string;
  caffeine: string;
  sources: string[];
  description: string;
  colorDescription: string;
  tasteDescription: string;
}